Create README.md
Manage Storage Optimization with Terminal: This project aims to optimize the storage of your Mac computer by using Terminal commands. The script performs various tasks such as deleting unnecessary files, clearing cache, and freeing up space. With this project, you can efficiently manage your storage and improve the performance of your Mac.

It is important to note that the process of managing storage optimization through the Terminal may cause issues for individuals who are not familiar with technical information.